# Openapi-Server-Stubs/OpenapiGenerator.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore",
    "version" : "1.0.5"
  },
...
  "servers" : [ {
    "url" : "https://petstore.swagger.io/"
  }, {
    "url" : "http://petstore.swagger.io/"
  } ],
...
  "paths" : {
...
    "/store/order" : {
      "post" : {
        "operationId" : "placeOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            }
          },
          "description" : "order placed for purchasing the pet",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid Order"
          }
        },
        "summary" : "Place an order for a pet",
        "tags" : [ "store" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/store/order/{orderId}" : {
      "delete" : {
        "description" : "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors",
        "operationId" : "deleteOrder",
        "parameters" : [ {
          "description" : "ID of the order that needs to be deleted",
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 1,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "400" : {
            "content" : { },
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "content" : { },
            "description" : "Order not found"
          }
        },
        "summary" : "Delete purchase order by ID",
        "tags" : [ "store" ]
      },
      "get" : {
        "description" : "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
        "operationId" : "getOrderById",
        "parameters" : [ {
          "description" : "ID of pet that needs to be fetched",
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "maximum" : 10,
            "minimum" : 1,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "content" : { },
            "description" : "Order not found"
          }
        },
        "summary" : "Find purchase order by ID",
        "tags" : [ "store" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Order" : {
        "example" : {
          "petId" : 6,
          "quantity" : 1,
          "id" : 0,
          "shipDate" : "2000-01-23T04:56:07.000+00:00",
          "complete" : true,
          "status" : "placed"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "petId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "shipDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "description" : "Order Status",
            "enum" : [ "placed", "approved", "delivered" ],
            "type" : "string"
          },
          "complete" : {
            "type" : "boolean"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "Order"
        }
      },
...
    }
  },
...
}